Adds some fun with structs and functs.
And, a fun little quirk with segmentation faults. See program.c for the two
pairs of lines of code that will and will not cause a segfault in, what I
assume, occurs in the construct() call. But... the newline `\n` is what is
causing it in main()... so.. that's an um.. interesting thing?